requirements.txt
fastmcp
elevenlabs
vercel_blob
python-dotenv